base-c/requirements.txt
alembic==1.2.1
asn1crypto==1.1.0
cffi==1.12.3
Click==7.0
cryptography==2.7
dateutils==0.6.6
elasticsearch==7.0.5
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==3.2.1
marshmallow-sqlalchemy==0.19.0
pycparser==2.19
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.3
redis==3.3.11
singletons==0.2.3
six==1.12.0
SQLAlchemy==1.3.10
urllib3==1.25.6
Werkzeug==0.16.0
